bpf: avoid out-of-range warning
The check against RINGBUF_MAX_DATA_SZ is known to be unnecessary on
32-bit architectures, but clang points out the same is true on
some 64-bit architectures, depending on the page size:

kernel/bpf/ringbuf.c:162:24: error: result of comparison of constant 1099511431168 with expression of type '__u32' (aka 'unsigned int') is always false [-Werror,-Wtautological-constant-out-of-range-compare]
        if (attr->max_entries > RINGBUF_MAX_DATA_SZ)
            ~~~~~~~~~~~~~~~~~ ^ ~~~~~~~~~~~~~~~~~~~

While correct, this warning is not helpful, as the compiler still does
the right thing for all configurations. Add a cast to shut up the
warning.

+++ kernel/bpf/ringbuf.c
@@ -159,7 +159,7 @@ static struct bpf_map *ringbuf_map_alloc(union bpf_attr *attr)
 
 #ifdef CONFIG_64BIT
 	/* on 32-bit arch, it's impossible to overflow record's hdr->pgoff */
-	if (attr->max_entries > RINGBUF_MAX_DATA_SZ)
+	if ((u64)attr->max_entries > RINGBUF_MAX_DATA_SZ)
 		return ERR_PTR(-E2BIG);
 #endif
